+++ packages/flodash/lib/_utils/_general.dart
@@ -1,0 +1,3 @@
+listEquals(List a, List b) {
+  return a.toSet().difference(b.toSet()).isEmpty && a.length == b.length;
+}