+++ packages/flodash/lib/modules/lang/is_equal.dart
@@ -1,7 +1,7 @@
-import 'package:flutter/foundation.dart';
+import '../../_utils/_general.dart';
 
 bool isEqual(a, b) {
   if (a is List && b is List) return listEquals(a, b);
-  if (a is Map && b is Map) return mapEquals(a, b);
+  if (a is Map && b is Map) return a.toString() == b.toString();
   return a == b;
 }