+++ packages/flodash/lib/modules/string/main.dart
@@ -9,3 +9,8 @@ export 'replace.dart';
 export 'upper_first.dart';
 export 'unescape.dart';
 export 'uppercase.dart';
+export 'trim_end.dart';
+export 'trim_start.dart';
+export 'starts_with.dart';
+export 'to_lower.dart';
+export 'to_upper.dart';