+++ packages/flodash/lib/modules/string/to_lower.dart
@@ -1,0 +1,15 @@
+/// Converts a string, as a whole, to lowercase.
+///
+/// Arguments
+///
+/// - (string): The string to convert.
+///
+/// Returns
+///
+/// - (string): The lowercase string.
+
+@Deprecated(
+    'The function toLower() is deprecated. Use the built-in toLowerCase() method instead.')
+String toLower([String string = '']) {
+  return string.toLowerCase();
+}