+++ packages/flodash/lib/modules/string/to_upper.dart
@@ -1,0 +1,15 @@
+/// Converts a string, as a whole, to uppercase.
+///
+/// Arguments
+///
+/// - (string): The string to convert.
+///
+/// Returns
+///
+/// - (string): The uppercase string.
+
+@Deprecated(
+    'The function toUpper() is deprecated. Use the built-in toUpperCase() method instead.')
+String toUpper([String string = '']) {
+  return string.toUpperCase();
+}