+++ packages/flodash/test/modules/array/chunk_test.dart
@@ -3,7 +3,7 @@ import 'package:flodash/flodash.dart' as flodash;
 import 'package:flutter_test/flutter_test.dart';
 
 main() {
-  List li = ["a", "b", "c", "d"];
+  List<String> li = ["a", "b", "c", "d"];
   test('even-split', () {
     expect(
         flodash.chunk(li, size: 2),