+++ packages/enhanced_cloud_firestore/CHANGELOG.md
@@ -1,3 +1,24 @@
+## 2.0.0-blizzard.2
+
+> Note: This release has breaking changes.
+
+ - **REFACTOR**: changes added.
+ - **BREAKING** **FEAT**: bumped firestore dependency versions.
+
+## 2.0.0-blizzard.1
+
+> Note: This release has breaking changes.
+
+ - **REFACTOR**: changes added.
+ - **BREAKING** **FEAT**: bumped firestore dependency versions.
+
+## 2.0.0-blizzard.0
+
+> Note: This release has breaking changes.
+
+ - **REFACTOR**: changes added.
+ - **BREAKING** **FEAT**: bumped firestore dependency versions.
+
 ## 1.0.1
 
 * Initial release.