+++ packages/flodash/CHANGELOG.md
@@ -1,3 +1,71 @@
+## 1.1.0-blizzard.6
+
+ - **FEAT**: added number clamp.
+
+## 1.1.0-blizzard.5
+
+> Note: This release has breaking changes.
+
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **FIX**: removed flutter dependency from flodash.
+ - **FEAT**: added starts_with under the string manipulation.
+ - **FEAT**: added to_lower under the string manipulation.
+ - **FEAT**: added to_upper under the string manipulation.
+ - **FEAT**: array/compact generic support.
+ - **FEAT**: added generics.
+ - **BREAKING** **FEAT**: bumped firestore dependency versions.
+
+## 1.1.0-blizzard.4
+
+> Note: This release has breaking changes.
+
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **FIX**: removed flutter dependency from flodash.
+ - **FEAT**: added starts_with under the string manipulation.
+ - **FEAT**: added to_lower under the string manipulation.
+ - **FEAT**: added to_upper under the string manipulation.
+ - **FEAT**: array/compact generic support.
+ - **FEAT**: added generics.
+ - **BREAKING** **FEAT**: bumped firestore dependency versions.
+
+## 1.1.0-blizzard.3
+
+> Note: This release has breaking changes.
+
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **REFACTOR**: changes added.
+ - **FEAT**: added starts_with under the string manipulation.
+ - **FEAT**: added to_lower under the string manipulation.
+ - **FEAT**: added to_upper under the string manipulation.
+ - **BREAKING** **FEAT**: bumped firestore dependency versions.
+
+## 1.1.0-blizzard.2
+
+ - **FEAT**: array/compact generic support.
+ - **FEAT**: added generics.
+
+## 1.1.0-blizzard.1
+
+ - **FEAT**: array/compact generic support.
+ - **FEAT**: added generics.
+
+## 1.1.0-blizzard.0
+
+ - **FEAT**: added generics.
+
+## 1.0.3
+
 ## 1.0.2
 
 * Fixed issue in number - range return type